requirements.txt
click==7.1.2
dnspython==2.1.0
Faker==8.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.15.0
text-unidecode==1.3
Werkzeug==1.0.1
